+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 1.1.5
+
+* 增加Cancellable中的WeakSet的初始化，使用时才进行初始化优化内存占用
+* 调整Cancellable.bindCancellable时throwWhenCancel=true，部分情况未抛出异常
+* Stream的扩展统一使用bindCancellable
+
 ## 1.1.4
 
 * Future.bindCancellable时允许以抛出异常[CancelledException]的方式继续